move data reshape functions
