+++ tests/risks/test_effect.py
@@ -5,7 +5,7 @@ from scipy.stats import norm
 
 from vivarium.framework.utilities import from_yearly
 from vivarium.testing_utilities import build_table, TestPopulation, metadata
-from vivarium.interface.interactive import setup_simulation, initialize_simulation
+from vivarium.interface.interactive import initialize_simulation
 
 from vivarium_public_health.disease import RateTransition
 from vivarium_public_health.risks.effect import DirectEffect, get_exposure_effect, rebin_rr_data
